Crate pallet_ethereum_checked

Ethereum Checked Pallet

Overview

A pallet-ethereum like pallet that execute transactions from checked source, like XCM remote call. Only Call transactions are supported (no Create).

The checked source guarantees that transactions are valid with prior checks, so these transactions are not required to include valid signatures. Instead, pallet-ethereum-checked will add the same dummy signature to them. To avoid transaction hash collisions, a global nonce shared with all users are used.

Interface

Dispatch-able calls

transact: transact an Ethereum transaction. Similar to pallet_ethereum::Transact, but is only for XCM remote call.

Re-exports

pub use weights::WeightInfo;

pub use pallet::*;

Modules

pallet

The pallet module in each FRAME pallet hosts the most important items needed to construct this pallet.

weights

Autogenerated weights for pallet_ethereum_checked

Structs

EnsureXcmEthereumTx

Ensure the origin is with XCM calls.

Enums

CheckedEthereumTxKind

Transaction kind.

RawOrigin

Origin for dispatch-able calls.

Type Aliases

WeightInfoOf