Trait pallet_ethereum_checked::pallet::Config

pub trait Config: Config + Config {
    type ReservedXcmpWeight: Get<Weight>;
    type XvmTxWeightLimit: Get<Weight>;
    type InvalidEvmTransactionError: From<TransactionValidationError>;
    type ValidatedTransaction: ValidatedTransaction;
    type AddressMapper: UnifiedAddressMapper<Self::AccountId>;
    type XcmTransactOrigin: EnsureOrigin<Self::RuntimeOrigin, Success = Self::AccountId>;
    type WeightInfo: WeightInfo;
}

Configuration trait of this pallet.

The main purpose of this trait is to act as an interface between this pallet and the runtime in which it is embedded in. A type, function, or constant in this trait is essentially left to be configured by the runtime that includes this pallet.

Consequently, a runtime that wants to include this pallet must implement this trait.

Required Associated Types

type ReservedXcmpWeight: Get<Weight>

Reserved Xcmp weight for block gas limit calculation.

type XvmTxWeightLimit: Get<Weight>

Xcm transaction weight limit, for block gas limit calculation.

type InvalidEvmTransactionError: From<TransactionValidationError>

Invalid tx error.

type ValidatedTransaction: ValidatedTransaction

Validated tx execution.

type AddressMapper: UnifiedAddressMapper<Self::AccountId>

Account mapping.

type XcmTransactOrigin: EnsureOrigin<Self::RuntimeOrigin, Success = Self::AccountId>

Origin for transact call.

type WeightInfo: WeightInfo

Weight information for extrinsics in this pallet.